soldunga: flavor text fixes, added flavor text for newly implemented classic npcs

+++ soldunga/young_goblin_wizard.lua
@@ -1,4 +1,5 @@
 function event_combat(e)
-	e.self:Say("Die by lava - Die by flame - Fire Peak goblins kill and maim.");
-
+	if(e.joined) then
+		e.self:Say("Die by lava - Die by flame - Fire Peak goblins kill and maim.");
+	end
 end
